socket updated

+++ BallAndPlate/Source/BallAndPlate/BallAndPlatePawn.h
@@ -29,7 +29,7 @@ public:
 
 	void setUpLights();
 
-	inline float map(long x, long in_min, long in_max, long out_min, long out_max)
+	inline float map(float x, float in_min, float in_max, float out_min, float out_max)
 	{
 		return (float)(x - in_min) * (out_max - out_min) / (float)(in_max - in_min) + out_min;
 	}
